requirements.txt
Django==3.1.4
djangorestframework==3.12.2
django-url-filter==0.3.15
django-cors-headers==3.6.0
django-heroku==0.3.1
dj-database-url==0.5.0
drf-yasg==1.20.0
gunicorn==20.0.4
python-dotenv==0.15.0
pre-commit==2.9.3
textdistance==4.2.0
psycopg2==2.8.6
django-modeltranslation==0.16.1
django-autoslug==1.9.8